feat(multi): implement G3 POS features, G7 dashboard comparison, G8 tax overrides, G9 P&L, G13 budgets, G14 backup scheduling, G10 cross-store customers
Features implemented across 24 files:

- G3-1: Cross-store return processing (POS state/intent/viewmodel)
- G3-3: Card terminal EMV connection status tracking
- G3-4: Wallet payment choice dialog with balance cap
- G7-1: Dashboard period comparison (yesterday + last week same-day)
- G8-1: Multi-region per-store tax overrides (Settings)
- G9-1: Multi-store P&L consolidation (Accounting)
- G10-1: Cross-store customer profile with per-store order summaries
- G13-1: Expense category budget tracking + monthly spend computation
- G13-2: Configurable expense approval threshold
- G14-1: Backup scheduling (frequency, hour, retention settings)
- fix(staff): Long→Int type mismatch in countDays

+++ composeApp/feature/staff/src/commonMain/kotlin/com/zyntasolutions/zyntapos/feature/staff/StaffViewModel.kt
@@ -753,7 +753,7 @@ class StaffViewModel(
             val start = kotlinx.datetime.LocalDate.parse(startDate)
             val end = kotlinx.datetime.LocalDate.parse(endDate)
             val diff = end.toEpochDays() - start.toEpochDays()
-            (diff + 1).coerceAtLeast(1)
+            (diff + 1).coerceAtLeast(1).toInt()
         } catch (_: Exception) {
             1 // Fallback: assume 1 day if dates can't be parsed
         }
